Generalize tail traversal logic in TailValueExtractor
